Modify the return value type of a single audio stream.

+++ frameworks/native/audiocapturer/include/audio_capturer_proxy_obj.h
@@ -31,7 +31,7 @@ public:
 
     void SetLowPowerVolumeImpl(float volume);
     void GetLowPowerVolumeImpl(float &volume);
-    void GetSingleStreamVolumeImpl(uint32_t &volume);
+    void GetSingleStreamVolumeImpl(float &volume);
 
 private:
     const AudioCapturer *capturer;
