+++ frameworks/native/audiostream/include/audio_stream.h
@@ -62,7 +62,7 @@ public:
     int32_t Clear();
     int32_t SetLowPowerVolume(float volume);
     float GetLowPowerVolume();
-    uint32_t GetSingleStreamVolume();
+    float GetSingleStreamVolume();
 
     std::vector<AudioSampleFormat> GetSupportedFormats() const;
     std::vector<AudioEncodingType> GetSupportedEncodingTypes() const;
