+++ services/include/audio_policy/client/audio_policy_base.h
@@ -38,7 +38,7 @@ public:
 
     virtual float GetLowPowerVolume(int32_t streamId) = 0;
 
-    virtual uint32_t GetSingleStreamVolume(int32_t streamId) = 0;
+    virtual float GetSingleStreamVolume(int32_t streamId) = 0;
 
     virtual int32_t SetStreamMute(AudioStreamType streamType, bool mute) = 0;
