+++ services/include/audio_policy/server/audio_client_tracker_callback_proxy.h
@@ -33,7 +33,7 @@ public:
 
     virtual void SetLowPowerVolumeImpl(float volume) override;
     virtual void GetLowPowerVolumeImpl(float &volume) override;
-    virtual void GetSingleStreamVolumeImpl(uint32_t &volume) override;
+    virtual void GetSingleStreamVolumeImpl(float &volume) override;
 
 private:
     sptr<IStandardClientTracker> listener_ = nullptr;
@@ -48,7 +48,7 @@ public:
 
     virtual void SetLowPowerVolumeImpl(float volume) override;
     virtual void GetLowPowerVolumeImpl(float &volume) override;
-    virtual void GetSingleStreamVolumeImpl(uint32_t &volume) override;
+    virtual void GetSingleStreamVolumeImpl(float &volume) override;
 
 private:
     static inline BrokerDelegator<AudioClientTrackerCallbackProxy> delegator_;
