+++ services/src/audio_policy/client/audio_policy_manager.cpp
@@ -122,7 +122,7 @@ float AudioPolicyManager::GetLowPowerVolume(int32_t streamId)
     return g_sProxy->GetLowPowerVolume(streamId);
 }
 
-uint32_t AudioPolicyManager::GetSingleStreamVolume(int32_t streamId)
+float AudioPolicyManager::GetSingleStreamVolume(int32_t streamId)
 {
     return g_sProxy->GetSingleStreamVolume(streamId);
 }
