+++ services/src/audio_policy/server/audio_stream_collector.cpp
@@ -491,11 +491,11 @@ float AudioStreamCollector::GetLowPowerVolume(int32_t streamId)
     return volume;
 }
 
-uint32_t AudioStreamCollector::GetSingleStreamVolume(int32_t streamId)
+float AudioStreamCollector::GetSingleStreamVolume(int32_t streamId)
 {
     CHECK_AND_RETURN_RET_LOG(!(clientTracker_.count(streamId) == 0),
         ERR_INVALID_PARAM, "AudioStreamCollector:GetSingleStreamVolume streamId invalid.");
-    uint32_t volume;
+    float volume;
     std::shared_ptr<AudioClientTracker> callback = clientTracker_[streamId];
     CHECK_AND_RETURN_RET_LOG(callback != nullptr,
         ERR_INVALID_PARAM, "AudioStreamCollector:GetSingleStreamVolume callback failed");
